+++ custom_components/tests/test_config_flow.py
@@ -1,4 +1,7 @@
+from copy import deepcopy
+import pytest
 import cololight
+
 
 from unittest.mock import patch
 
@@ -7,41 +10,42 @@ from tests.common import MockConfigEntry
 
 from homeassistant import data_entry_flow
 
-
 NAME = "cololight_test"
 HOST = "1.1.1.1"
-NAME_2 = "cololight_test_2"
-HOST_2 = "1.1.1.2"
 
 DEMO_USER_INPUT = {
-    "name": NAME,
-    "host": HOST,
-    "default_effects": [
-        "80s Club",
-        "Cherry Blossom",
-    ],
+    "device_data": {"name": NAME, "host": HOST, "device": "hexagon"},
+    "effects_data": {
+        "default_effects": [
+            "80s Club",
+            "Cherry Blossom",
+        ],
+    },
 }
 
 DEMO_USER_INPUT_2 = {
-    "name": NAME_2,
-    "host": HOST_2,
-    "default_effects": [
-        "80s Club",
-        "Cherry Blossom",
-    ],
-}
-
-DEMO_USER_INPUT_3 = {
-    "name": NAME,
-    "host": HOST,
-    "default_effects": [
-        "80s Club",
-        "Cherry Blossom",
-    ],
+    "device_data": {"name": NAME, "host": HOST, "device": "strip"},
+    "effects_data": {
+        "default_effects": [
+            "80s Club",
+            "Cherry Blossom",
+        ],
+        "dynamic_effects": ["Graffiti", "Snow"],
+    },
 }
 
 
-async def test_form(hass):
+@pytest.fixture
+def demo_user_input():
+    return deepcopy(DEMO_USER_INPUT)
+
+
+@pytest.fixture
+def demo_user_input_2():
+    return deepcopy(DEMO_USER_INPUT_2)
+
+
+async def test_form(hass, demo_user_input):
     """Test config entry configured successfully."""
 
     result = await hass.config_entries.flow.async_init(
@@ -51,25 +55,77 @@ async def test_form(hass):
     assert result["step_id"] == "user"
 
     result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], user_input=DEMO_USER_INPUT
+        result["flow_id"], user_input=demo_user_input["device_data"]
     )
+
+    assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
+    assert result["step_id"] == "device_effects"
+
+    result = await hass.config_entries.flow.async_configure(
+        result["flow_id"],
+        user_input=demo_user_input["effects_data"],
+    )
+
+    expected_data = demo_user_input["device_data"] | demo_user_input["effects_data"]
 
     assert result["type"] == "create_entry"
     assert result["title"] == NAME
-    assert result["data"] == DEMO_USER_INPUT
+    assert result["data"] == expected_data
+
+    await hass.async_block_till_done()
+    assert hass.data["cololight"][result["result"].entry_id].effects == [
+        "80s Club",
+        "Cherry Blossom",
+    ]
 
 
-async def test_form_already_configured(hass):
+async def test_form_strip(hass, demo_user_input_2):
+    """Test config entry configured successfully."""
+
+    result = await hass.config_entries.flow.async_init(
+        cololight.DOMAIN, context={"source": "user"}
+    )
+    assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
+    assert result["step_id"] == "user"
+
+    result = await hass.config_entries.flow.async_configure(
+        result["flow_id"], user_input=demo_user_input_2["device_data"]
+    )
+
+    assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
+    assert result["step_id"] == "device_effects"
+
+    result = await hass.config_entries.flow.async_configure(
+        result["flow_id"],
+        user_input=demo_user_input_2["effects_data"],
+    )
+
+    expected_data = demo_user_input_2["device_data"] | demo_user_input_2["effects_data"]
+
+    assert result["type"] == "create_entry"
+    assert result["title"] == NAME
+    assert result["data"] == expected_data
+
+    await hass.async_block_till_done()
+    assert hass.data["cololight"][result["result"].entry_id].effects == [
+        "80s Club",
+        "Cherry Blossom",
+        "Graffiti",
+        "Snow",
+    ]
+
+
+async def test_form_already_configured(hass, demo_user_input):
     """Test host is already configured."""
     entry = MockConfigEntry(
-        domain=cololight.DOMAIN, data=DEMO_USER_INPUT, unique_id=HOST
+        domain=cololight.DOMAIN, data=demo_user_input["device_data"], unique_id=HOST
     )
     entry.add_to_hass(hass)
 
     result = await hass.config_entries.flow.async_init(
         cololight.DOMAIN,
         context={"source": "user"},
-        data=DEMO_USER_INPUT,
+        data=demo_user_input["device_data"],
     )
 
     assert result["type"] == "abort"
@@ -80,24 +136,33 @@ async def test_form_already_configured(hass):
     "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_color_schemes",
     return_value=["Mood | Green", "Mood | Red"],
 )
-async def test_options_creating_effect(mock_color_schemes, hass):
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_cololight"
+)
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_max_mode",
+    return_value=27,
+)
+async def test_options_creating_effect(
+    mock_max_mode, mock_cololight, mock_color_schemes, hass, demo_user_input
+):
     """Test options for create effect"""
     entry = MockConfigEntry(
-        domain=cololight.DOMAIN, data=DEMO_USER_INPUT, unique_id=HOST
+        domain=cololight.DOMAIN, data=demo_user_input, unique_id=HOST
     )
     entry.add_to_hass(hass)
 
     result = await hass.config_entries.options.async_init(entry.entry_id)
 
-    assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
+    assert result["type"] == data_entry_flow.RESULT_TYPE_MENU
     assert result["step_id"] == "init"
 
     result = await hass.config_entries.options.async_configure(
-        result["flow_id"], user_input={"select": "Create"}
+        result["flow_id"], user_input={"next_step_id": "create_effect"}
     )
 
     assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
-    assert result["step_id"] == "options_add_custom_effect"
+    assert result["step_id"] == "create_effect"
 
     result = await hass.config_entries.options.async_configure(
         result["flow_id"],
@@ -122,7 +187,7 @@ async def test_options_creating_effect(mock_color_schemes, hass):
     result = await hass.config_entries.options.async_init(entry.entry_id)
 
     result = await hass.config_entries.options.async_configure(
-        result["flow_id"], user_input={"select": "Create"}
+        result["flow_id"], user_input={"next_step_id": "create_effect"}
     )
 
     result = await hass.config_entries.options.async_configure(
@@ -155,16 +220,25 @@ async def test_options_creating_effect(mock_color_schemes, hass):
     "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_color_schemes",
     return_value=["Mood | Green", "Mood | Red"],
 )
-async def test_options_updating_effect(mock_scheme_colors, hass):
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_cololight"
+)
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_max_mode",
+    return_value=27,
+)
+async def test_options_updating_effect(
+    mock_max_mode, mock_cololight, mock_scheme_colors, hass, demo_user_input
+):
     """Test options for create effect is updated if already existing"""
     entry = MockConfigEntry(
-        domain=cololight.DOMAIN, data=DEMO_USER_INPUT, unique_id=HOST
+        domain=cololight.DOMAIN, data=demo_user_input, unique_id=HOST
     )
     entry.add_to_hass(hass)
 
     result = await hass.config_entries.options.async_init(entry.entry_id)
     result = await hass.config_entries.options.async_configure(
-        result["flow_id"], user_input={"select": "Create"}
+        result["flow_id"], user_input={"next_step_id": "create_effect"}
     )
     result = await hass.config_entries.options.async_configure(
         result["flow_id"],
@@ -186,7 +260,7 @@ async def test_options_updating_effect(mock_scheme_colors, hass):
 
     result = await hass.config_entries.options.async_init(entry.entry_id)
     result = await hass.config_entries.options.async_configure(
-        result["flow_id"], user_input={"select": "Create"}
+        result["flow_id"], user_input={"next_step_id": "create_effect"}
     )
 
     result = await hass.config_entries.options.async_configure(
@@ -218,8 +292,13 @@ async def test_options_updating_effect(mock_scheme_colors, hass):
         "test_2": "test_2",
     },
 )
-async def test_options_deleting_custom_effect(mock_effects, hass):
-    """Test options for deleting effect"""
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_cololight"
+)
+async def test_options_removing_custom_effect(
+    mock_cololight, mock_effects, hass, demo_user_input
+):
+    """Test options for removing effect"""
     test_effects = {
         "test": {"color_scheme": "Mood", "color": "Green", "cycle_speed": 1, "mode": 1},
         "test_2": {
@@ -240,7 +319,7 @@ async def test_options_deleting_custom_effect(mock_effects, hass):
     }
 
     entry = MockConfigEntry(
-        domain=cololight.DOMAIN, data=DEMO_USER_INPUT, unique_id=HOST
+        domain=cololight.DOMAIN, data=demo_user_input, unique_id=HOST
     )
     entry.add_to_hass(hass)
 
@@ -248,15 +327,15 @@ async def test_options_deleting_custom_effect(mock_effects, hass):
 
     result = await hass.config_entries.options.async_init(entry.entry_id)
 
-    assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
+    assert result["type"] == data_entry_flow.RESULT_TYPE_MENU
     assert result["step_id"] == "init"
 
     result = await hass.config_entries.options.async_configure(
-        result["flow_id"], user_input={"select": "Delete"}
+        result["flow_id"], user_input={"next_step_id": "remove_effect"}
     )
 
     assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
-    assert result["step_id"] == "options_delete_effect"
+    assert result["step_id"] == "remove_effect"
 
     result = await hass.config_entries.options.async_configure(
         result["flow_id"],
@@ -265,17 +344,24 @@ async def test_options_deleting_custom_effect(mock_effects, hass):
         },
     )
 
+    await hass.async_block_till_done()
+
     assert result["type"] == data_entry_flow.RESULT_TYPE_CREATE_ENTRY
     assert entry.options == expected_effetcs
 
 
-async def test_options_deleting_default_effect(hass):
-    """Test options for deleting effect"""
+async def test_options_removing_default_effect(hass, demo_user_input):
+    """Test options for removing effect"""
     flow = await hass.config_entries.flow.async_init(
         cololight.DOMAIN, context={"source": "user"}
     )
     entry = await hass.config_entries.flow.async_configure(
-        flow["flow_id"], user_input=DEMO_USER_INPUT
+        flow["flow_id"], user_input=demo_user_input["device_data"]
+    )
+
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"],
+        user_input=demo_user_input["effects_data"],
     )
 
     await hass.async_block_till_done()
@@ -283,7 +369,7 @@ async def test_options_deleting_default_effect(hass):
     option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
 
     option_flow = await hass.config_entries.options.async_configure(
-        option_flow["flow_id"], user_input={"select": "Delete"}
+        option_flow["flow_id"], user_input={"next_step_id": "remove_effect"}
     )
 
     await hass.config_entries.options.async_configure(
@@ -297,13 +383,18 @@ async def test_options_deleting_default_effect(hass):
     assert hass.data["cololight"][entry["result"].entry_id].effects == ["80s Club"]
 
 
-async def test_options_restoring_default_effect(hass):
-    """Test options for restoring effect"""
+async def test_options_removing_dynamic_effect(hass, demo_user_input_2):
+    """Test options for removing effect"""
     flow = await hass.config_entries.flow.async_init(
         cololight.DOMAIN, context={"source": "user"}
     )
     entry = await hass.config_entries.flow.async_configure(
-        flow["flow_id"], user_input=DEMO_USER_INPUT_3
+        flow["flow_id"], user_input=demo_user_input_2["device_data"]
+    )
+
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"],
+        user_input=demo_user_input_2["effects_data"],
     )
 
     await hass.async_block_till_done()
@@ -311,13 +402,50 @@ async def test_options_restoring_default_effect(hass):
     option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
 
     option_flow = await hass.config_entries.options.async_configure(
-        option_flow["flow_id"], user_input={"select": "Restore"}
+        option_flow["flow_id"], user_input={"next_step_id": "remove_effect"}
     )
 
     await hass.config_entries.options.async_configure(
         option_flow["flow_id"],
         user_input={
-            "name": ["Unicorns"],
+            "name": ["Graffiti"],
+        },
+    )
+
+    await hass.async_block_till_done()
+    assert hass.data["cololight"][entry["result"].entry_id].effects == [
+        "80s Club",
+        "Cherry Blossom",
+        "Snow",
+    ]
+
+
+async def test_options_restoring_default_effect(hass, demo_user_input):
+    """Test options for restoring effect"""
+    flow = await hass.config_entries.flow.async_init(
+        cololight.DOMAIN, context={"source": "user"}
+    )
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"], user_input=demo_user_input["device_data"]
+    )
+
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"],
+        user_input=demo_user_input["effects_data"],
+    )
+
+    await hass.async_block_till_done()
+
+    option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
+
+    option_flow = await hass.config_entries.options.async_configure(
+        option_flow["flow_id"], user_input={"next_step_id": "restore_effect"}
+    )
+
+    await hass.config_entries.options.async_configure(
+        option_flow["flow_id"],
+        user_input={
+            "default_effects": ["Unicorns"],
         },
     )
 
@@ -329,28 +457,74 @@ async def test_options_restoring_default_effect(hass):
     ]
 
 
+async def test_options_restoring_dynamic_effect(hass, demo_user_input_2):
+    """Test options for restoring effect"""
+    flow = await hass.config_entries.flow.async_init(
+        cololight.DOMAIN, context={"source": "user"}
+    )
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"], user_input=demo_user_input_2["device_data"]
+    )
+
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"],
+        user_input=demo_user_input_2["effects_data"],
+    )
+
+    await hass.async_block_till_done()
+
+    option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
+
+    option_flow = await hass.config_entries.options.async_configure(
+        option_flow["flow_id"], user_input={"next_step_id": "restore_effect"}
+    )
+
+    await hass.config_entries.options.async_configure(
+        option_flow["flow_id"],
+        user_input={"dynamic_effects": ["Color train"]},
+    )
+
+    await hass.async_block_till_done()
+    assert hass.data["cololight"][entry["result"].entry_id].effects == [
+        "80s Club",
+        "Cherry Blossom",
+        "Graffiti",
+        "Snow",
+        "Color train",
+    ]
+
+
 @patch(
     "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_color_schemes",
     return_value=["Mood | Green"],
 )
-async def test_options_has_error_if_invalid(mock_color_schemes, hass):
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_cololight"
+)
+@patch(
+    "homeassistant.components.cololight.config_flow.CololightOptionsFlowHandler._get_max_mode",
+    return_value=27,
+)
+async def test_options_has_error_if_invalid(
+    mock_max_mode, mock_cololight, mock_color_schemes, hass, demo_user_input
+):
     """Test options will show error if invalid"""
     entry = MockConfigEntry(
-        domain=cololight.DOMAIN, data=DEMO_USER_INPUT, unique_id=HOST
+        domain=cololight.DOMAIN, data=demo_user_input, unique_id=HOST
     )
     entry.add_to_hass(hass)
 
     result = await hass.config_entries.options.async_init(entry.entry_id)
 
-    assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
+    assert result["type"] == data_entry_flow.RESULT_TYPE_MENU
     assert result["step_id"] == "init"
 
     result = await hass.config_entries.options.async_configure(
-        result["flow_id"], user_input={"select": "Create"}
+        result["flow_id"], user_input={"next_step_id": "create_effect"}
     )
 
     assert result["type"] == data_entry_flow.RESULT_TYPE_FORM
-    assert result["step_id"] == "options_add_custom_effect"
+    assert result["step_id"] == "create_effect"
 
     result = await hass.config_entries.options.async_configure(
         result["flow_id"],
@@ -379,13 +553,18 @@ async def test_options_has_error_if_invalid(mock_color_schemes, hass):
     assert result["errors"] == {"mode": "invalid_mode"}
 
 
-async def test_end_to_end_with_options(hass):
-    """Test an end to end flow, creating entity and add effects and deleting effects"""
+async def test_end_to_end_with_options(hass, demo_user_input):
+    """Test an end to end flow, creating entity and add effects and removing effects"""
     flow = await hass.config_entries.flow.async_init(
         cololight.DOMAIN, context={"source": "user"}
     )
     entry = await hass.config_entries.flow.async_configure(
-        flow["flow_id"], user_input=DEMO_USER_INPUT_2
+        flow["flow_id"], user_input=demo_user_input["device_data"]
+    )
+
+    entry = await hass.config_entries.flow.async_configure(
+        flow["flow_id"],
+        user_input=demo_user_input["effects_data"],
     )
 
     await hass.async_block_till_done()
@@ -393,7 +572,7 @@ async def test_end_to_end_with_options(hass):
     option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
 
     option_flow = await hass.config_entries.options.async_configure(
-        option_flow["flow_id"], user_input={"select": "Create"}
+        option_flow["flow_id"], user_input={"next_step_id": "create_effect"}
     )
 
     await hass.config_entries.options.async_configure(
@@ -412,7 +591,7 @@ async def test_end_to_end_with_options(hass):
     option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
 
     option_flow = await hass.config_entries.options.async_configure(
-        option_flow["flow_id"], user_input={"select": "Delete"}
+        option_flow["flow_id"], user_input={"next_step_id": "remove_effect"}
     )
 
     await hass.config_entries.options.async_configure(
@@ -430,7 +609,7 @@ async def test_end_to_end_with_options(hass):
     option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
 
     option_flow = await hass.config_entries.options.async_configure(
-        option_flow["flow_id"], user_input={"select": "Delete"}
+        option_flow["flow_id"], user_input={"next_step_id": "remove_effect"}
     )
 
     await hass.config_entries.options.async_configure(
@@ -446,13 +625,13 @@ async def test_end_to_end_with_options(hass):
     option_flow = await hass.config_entries.options.async_init(entry["result"].entry_id)
 
     option_flow = await hass.config_entries.options.async_configure(
-        option_flow["flow_id"], user_input={"select": "Restore"}
+        option_flow["flow_id"], user_input={"next_step_id": "restore_effect"}
     )
 
     await hass.config_entries.options.async_configure(
         option_flow["flow_id"],
         user_input={
-            "name": ["Savasana", "Unicorns"],
+            "default_effects": ["Savasana", "Unicorns"],
         },
     )